+++ app/documentation/index.js
@@ -7,9 +7,12 @@ const cache = new Cache(config.cache_directory, { minify: true, gzip: true });
 const fs = require("fs-extra");
 const redirector = require("./redirector");
 const trace = require("helper/trace");
+const hash = require("helper/hash");
+const { blot_directory } = require("config");
 
 const root = require("helper/rootDir");
 const { join } = require("path");
+const { id } = require("typeset/src/hypher-patterns/en-us");
 const VIEW_DIRECTORY = join(root, "app/views");
 
 // Register the engine we will use to
@@ -54,8 +57,30 @@ documentation.locals.interval = plan.startsWith("monthly") ? "month" : "year";
 
 let cacheID = Date.now();
 
-documentation.locals.cdn = () => (text, render) =>
-  `/cdn/documentation/${cacheID}${render(text)}`;
+documentation.locals.cdn = () => (text, render) => {
+  const path = render(text);
+  const extension = path.split(".").pop();
+
+  let identifier = "cacheID=" + cacheID;
+
+  try {
+    const contents = fs.readFileSync(
+      join(blot_directory, "/static/documentation", path),
+      "utf8"
+    );
+
+    const hash = hash(contents);
+
+    identifier = "hash=" + hash;
+  } catch (e) {
+    // if the file doesn't exist, we'll use the cacheID
+  }
+
+  const query = `?${identifier}&ext=${extension}`;
+  const url = `${config.cdn.origin}/documentation${path}${query}`;
+
+  return url;
+};
 
 if (config.environment === "development") {
   const chokidar = require("chokidar");
