+++ config/nginx/blot-cdn.conf
@@ -31,7 +31,6 @@ location / {
   add_header 'Blot-Cache' 'true-cdn-new' always;
   add_header 'Cache-Control' 'public, max-age=31536000' always;
   add_header 'Access-Control-Allow-Origin' '*';
-  client_max_body_size 1M;
   gzip on;
   try_files $static_directory $global_static_directory =404;
 }
@@ -42,6 +41,7 @@ location = / {
   return 302 https://$blot_host;
 }
 
+# allow search engines to index the site
 location = /robots.txt {
   return 200 "User-agent: *\nAllow: /\n";
 }
@@ -51,6 +51,7 @@ location = /favicon.ico {
     root $blot_directory/app/views;    
 }
 
+# for missing files
 location = /error-404.html {
     root  $blot_directory/app/views;
 }
