+++ config/nginx/blot-site.conf
@@ -69,26 +69,6 @@ location ^~ /.git {
   include /var/www/blot/config/nginx/reverse-proxy.conf;
 }
 
-# location @permanent_cache_gzip {
-#   add_header 'Blot-Cache' 'true-permanent-gzip' always;
-#   add_header 'Cache-Control' 'public, max-age=31536000' always;
-#   add_header 'Access-Control-Allow-Origin' '*';
-#   add_header 'Content-Encoding' 'gzip';
-#   client_max_body_size 1M;
-#   # We have pre-compressed this file if it exists
-#   gzip off;
-#   try_files $permanent_cache$gz @permanent_cache;  
-# }
-
-location @permanent_cache {
-  add_header 'Blot-Cache' 'true-permanent' always;
-  add_header 'Cache-Control' 'public, max-age=31536000' always;
-  add_header 'Access-Control-Allow-Origin' '*';
-  client_max_body_size 1M;
-  gzip on;
-  try_files $permanent_cache @temporary_cache;  
-}
-
 
 # This is where we send server-sent events
 # which need a long timeout
@@ -122,14 +102,12 @@ location /clients/git/end/ {
 }
 
 location / {
-  add_header 'Blot-Cache' 'true-temporary-gzip' always;
-  add_header 'Cache-Control' 'no-cache' always;
-  add_header 'Content-Encoding' 'gzip';
+  add_header 'Blot-Cache' 'true-permanent' always;
+  add_header 'Cache-Control' 'public, max-age=31536000' always;
+  add_header 'Access-Control-Allow-Origin' '*';
   client_max_body_size 1M;
-  # We have pre-compressed this file if it exists
-  gzip off;
-  open_file_cache off;
-  try_files @permanent_cache @blot;  
+  gzip on;
+  try_files $permanent_cache @temporary_cache;  
 }
 
 location @temporary_cache {
@@ -138,7 +116,16 @@ location @temporary_cache {
   client_max_body_size 1M;
   gzip on;
   open_file_cache off;
-  try_files $temporary_cache @blot;  
+  try_files $temporary_cache @documentation_static_files;  
+}
+
+location @documentation_static_files {
+  add_header 'Blot-Cache' 'static-docs' always;
+  add_header 'Cache-Control' 'public, max-age=31536000' always;
+  add_header 'Access-Control-Allow-Origin' '*';
+  gzip on;
+  # $uri rather than $request_uri because we don't want the query string
+  try_files $blot_directory/app/views$uri @blot;
 }
 
 location @blot {
